requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.3.0
#certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==3.1.0
comm==0.1.1
decorator==5.1.1
defusedxml==0.7.1
efficientnet-pytorch==0.7.1
filelock==3.11.0
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.3
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.53.0
huggingface-hub==0.13.4
idna==3.4
imageio==2.27.0
imagezmq==1.1.1
importlib-metadata==6.3.0
imutils==0.5.4
install==1.3.5
ipython-genutils==0.2.0
ipywidgets==8.0.2
Jinja2==3.1.2
joblib==1.2.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
lazy_loader==0.2
Markdown==3.4.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.4
munch==2.5.0
networkx==3.1
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==21.3
pandas==2.0.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.5.4
pretrainedmodels==0.7.4
protobuf==4.22.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.1
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytorch-lightning==1.5.4
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
qudida==0.0.4
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
segmentation-models-pytorch==0.3.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.6.12
torch==2.0.1
torchaudio==2.0.2
torchmetrics==1.2.0
torchvision==0.15.2
tornado==6.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.3
yarl==1.8.2
zipp==3.15.0
